* edit swiped pictures on components

* applied apis to page beranda and virtual tour and change the footer

* branchKantor 7/10/24

* from branchRumah 07/10/24

* so far all done from branchKantor 08/10/24

* rating avg tofixed (1) , perubahan card jika ada chip, swipedpictures jika hp margin diubah

* perubahan navbar, review + penambahan login dan logout

* change header + review for getstorage (useeffect added)

---------

+++ utils/storage.js
@@ -1,15 +1,24 @@
 export const getStorage = (key) => {
-  return localStorage.getItem(key);
+  if (typeof window !== "undefined") {
+    return localStorage.getItem(key);
+  }
+  return null; // or a default value
 };
 
 export const setStorage = (key, val) => {
-  localStorage.setItem(key, val);
+  if (typeof window !== "undefined") {
+    localStorage.setItem(key, val);
+  }
 };
 
 export const deleteStorage = (key) => {
-  localStorage.removeItem(key);
+  if (typeof window !== "undefined") {
+    localStorage.removeItem(key);
+  }
 };
 
 export const clearStorage = () => {
-  localStorage.clear();
+  if (typeof window !== "undefined") {
+    localStorage.clear();
+  }
 };
